GSB: Correct the stats in GSB and introduce packet drop mechanism
To check the performance of Generic SW bridge, we introduced
this change to drop all packets before xmit and before giving
packets to IPA as well as NW stack. This change is controlled
by toggling a proc entry *gsb_drop_pkts* which is set to zero
initially.

Stats of packets fed back to GSB were recounted and we have
handled that through this change by checking the loop tag
in skb control buffer and not incrementing the stats.

Change-Id: Ib07bf0df395ee02965d9d5c301995ff19f49be4c
Acked-by: Rishav LNU <rna@qti.qualcomm.com>
Signed-off-by: Mohammed Javid <mjavid@codeaurora.org>