drivers: Add rmnet-nss module
The RmNet-NSS module provides a link between the RmNet and NSS drivers
to allow for packets arriving from modem to take advantage of hardware
acceleration capabilities provided by the NSS driver for Ethernet packets.

Packets coming from the modem are deaggregated and demuxed as normal by
the RmNet driver, and are then stamped with Ethernet headers and handed to
the NSS driver for accelerated processing when the module is loaded.
Similarly, the module will register with NSS to be have a callback invoked
allowing NSS to pass the RmNet driver packets intended for transmission
over the modem. The module will handle removing the Ethernet headers from
these packets and will pass them to the base RmNet driver so they can be
stamped with appropriate QMAP headers suitable for sending to the modem.

Change-Id: I2af5af4f80ecc35e6f877e5550fecde78cc467ef
Signed-off-by: Sean Tranchetti <stranche@codeaurora.org>