GSB: Better approach to identify the debugfs node name
Earlier, code was evaluating the complete raw path of debugfs dir
and processing the path to get debugfs node name. New approach is
to use the the data structures which carry the name information.
Additionally, debug apis are changes to address a condition where
new line char is displayed while prnting debug messages to dmesg.

Acked-by: Gajendra Singh <gajendra@qti.qualcomm.com>
Signed-off-by: Bojun Pan <bojunp@codeaurora.org>